LayoutTests/http/tests/websocket/tests/hybi/url-with-query.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Make sure handshake with URL with query components success.");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var url = "ws://127.0.0.1:8880/websocket/tests/hybi/echo-location?q=v";
var handshake_success = false;
var ws_location;

function endTest()
{
    clearTimeout(timeoutID);
    shouldBeTrue("handshake_success");
    shouldBe("ws_location", "url");
    finishJSTest();
}

debug("url=" + url);
var ws = new WebSocket(url);
ws.onopen = function () {
    debug("WebSocket is open");
    handshake_success = true;
};
ws.onmessage = function (evt) {
    ws_location = evt.data;
    debug("received:" + ws_location);
    ws.close();
};
ws.onclose = function () {
    debug("WebSocket is closed");
    endTest();
};
var timeoutID = setTimeout("endTest()", 2000);

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
