LayoutTests/http/tests/websocket/tests/hybi/url-with-query-expected.txt

Make sure handshake with URL with query components success.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

url=ws://127.0.0.1:8880/websocket/tests/hybi/echo-location?q=v
WebSocket is open
received:ws://127.0.0.1:8880/websocket/tests/hybi/echo-location?q=v
WebSocket is closed
PASS handshake_success is true
PASS ws_location is url
PASS successfullyParsed is true

TEST COMPLETE
